+++ client/src/features/tracks/index.js
@@ -1,23 +1,33 @@
 import axios from 'axios';
 import { createSlice } from '@reduxjs/toolkit';
 import { setActiveRelease } from 'features/releases';
-import { toastError } from 'features/toast';
+import { toastError, toastInfo, toastSuccess } from 'features/toast';
+const calls = new Map();
 
 const trackSlice = createSlice({
   name: 'tracks',
   initialState: {
     audioUploadProgress: {},
-    cancelAudioUpload: undefined,
-    isDeleting: [],
+    uploadCancelled: false,
+    trackIdsForDeletion: {},
     isTranscoding: []
   },
   reducers: {
-    setDeletingComplete(state, action) {
-      state.isDeleting = state.isDeleting.filter(id => id !== action.payload);
+    cancelUpload(state, action) {
+      const trackId = action.payload;
+      const call = calls.get(trackId);
+
+      if (call) {
+        call.cancel();
+        calls.delete(trackId);
+      }
+
+      state.uploadProgress = { ...state.audioUploadProgress, [trackId]: 0 };
     },
 
-    setDeletingStart(state, action) {
-      state.isDeleting = [...state.isDeleting, action.payload];
+    setTrackIdsForDeletion(state, action) {
+      const { trackId, isDeleting } = action.payload;
+      state.trackIdsForDeletion = { ...state.trackIdsForDeletion, [trackId]: isDeleting };
     },
 
     setTranscodingComplete(state, action) {
@@ -35,24 +45,34 @@ const trackSlice = createSlice({
   }
 });
 
-const addTrack = releaseId => async dispatch => {
+const addTrack = (releaseId, newTrack) => async dispatch => {
   try {
-    const res = await axios.put(`/api/track/${releaseId}/add`);
+    const res = await axios.put(`/api/track/${releaseId}/add`, { newTrack });
     dispatch(setActiveRelease(res.data));
   } catch (error) {
     dispatch(toastError(error.response.data.error));
   }
 };
 
-const deleteTrack = (releaseId, trackId) => async dispatch => {
+const deleteTrack = (releaseId, trackId, trackTitle) => async (dispatch, getState) => {
   try {
-    dispatch(setDeletingStart(trackId));
-    const res = await axios.delete(`/api/track/${releaseId}/${trackId}`);
-    dispatch(setActiveRelease(res.data));
-    dispatch(setDeletingComplete(trackId));
+    if (getState().tracks.trackIdsForDeletion[trackId]) {
+      const res = await axios.delete(`/api/track/${releaseId}/${trackId}`);
+      dispatch(setActiveRelease(res.data));
+      dispatch(toastSuccess(`${trackTitle ? `\u2018${trackTitle}\u2019` : 'Track'} deleted.`));
+      dispatch(setTrackIdsForDeletion({ trackId, isDeleting: false }));
+    } else {
+      dispatch(setTrackIdsForDeletion({ trackId, isDeleting: true }));
+    }
   } catch (error) {
     dispatch(toastError(error.response.data.error));
   }
+};
+
+const getCancelToken = uploadId => dispatch => {
+  const call = axios.CancelToken.source();
+  calls.set(uploadId, call);
+  return call.token;
 };
 
 const moveTrack = (releaseId, fromIndex, toIndex) => async dispatch => {
@@ -82,29 +102,36 @@ const uploadAudio = ({ releaseId, trackId, trackName, audioFile, type }) => asyn
     formData.append('trackId', trackId);
     formData.append('trackName', trackName);
     formData.append('file', audioFile, audioFile.name);
+    const cancelToken = dispatch(getCancelToken(trackId));
 
     const config = {
       onUploadProgress: event => {
         const percent = Math.floor((event.loaded / event.total) * 100);
         dispatch(setUploadProgress({ trackId, percent }));
-      }
+      },
+      cancelToken
     };
 
-    const res = await axios.post('/api/track/upload', formData, config);
-    return res;
+    await axios.post('/api/track/upload', formData, config);
   } catch (error) {
+    if (axios.isCancel(error)) {
+      return toastInfo('Upload cancelled.');
+    }
+
     toastError(error.response.data.error);
     dispatch(setUploadProgress({ trackId, percent: 0 }));
   }
 };
 
 export const {
+  cancelUpload,
   setDeletingComplete,
   setDeletingStart,
+  setTrackIdsForDeletion,
   setTranscodingComplete,
   setTranscodingStart,
   setUploadProgress
 } = trackSlice.actions;
 
-export { addTrack, deleteTrack, moveTrack, transcodeAudio, uploadAudio };
+export { addTrack, deleteTrack, getCancelToken, moveTrack, transcodeAudio, uploadAudio };
 export default trackSlice.reducer;
